backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
mysql-connector-python>=9.0.0
python-dotenv>=1.0.0
pydantic[email]>=2.0.0
PyJWT[crypto]>=2.8.0
bcrypt>=4.2.0
google-auth>=2.29.0
